2024.01.17 1st commit.
Added integration tests for Create, Update and Delete methods of BookController class.
Updated BookService.Delete and added a new unit test to cover update.

+++ Server/ReadingClub/Services/Implementations/BookService.cs
@@ -41,6 +41,11 @@ namespace ReadingClub.Services.Implementations
 
             var book = await _bookRepository.Get(id);
 
+            if (book == null)
+            {
+                return null!;
+            }
+
             book.AddedBy = anonymousUserId;
 
             var updatedBook = await _bookRepository.DeleteUpdate(book);
